+++ bundle/PolicyProvider/InformationCollectionPolicyProvider.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Netgen\Bundle\InformationCollectionBundle\PolicyProvider;
+
+use eZ\Bundle\EzPublishCoreBundle\DependencyInjection\Configuration\ConfigBuilderInterface;
+use eZ\Bundle\EzPublishCoreBundle\DependencyInjection\Security\PolicyProvider\PolicyProviderInterface;
+
+class InformationCollectionPolicyProvider implements PolicyProviderInterface
+{
+    /**
+     * @param \eZ\Bundle\EzPublishCoreBundle\DependencyInjection\Configuration\ConfigBuilderInterface $configBuilder
+     *
+     * @return array
+     */
+    public function addPolicies(ConfigBuilderInterface $configBuilder)
+    {
+        $configBuilder->addConfig([
+            'infocollector' => [
+                'read' => [],
+                'delete' => [],
+                'anonymize' => [],
+                'export' => [],
+            ],
+        ]);
+
+        return [];
+    }
+}